requirements-dev.txt
pytest
flake8
black
mypy
isort
types-setuptools
types-requests
pandas-stubs